cambios en el token y mejorando la funcionalidad de seguir
