+++ Frontend/src/Components/Suggestions/CardSuggestions.jsx
@@ -1,29 +1,22 @@
 import axios from "axios";
-import { useEffect, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useState } from "react";
+import { useSelector } from "react-redux";
 import { Link } from "react-router-dom";
-import { clearUsersState, getUsers } from "../../actions";
 import { tokenUsuario } from "../../actions/actionTypes";
 import { StyledCardSuggestions } from "./StyledCardSuggestion";
 
 const CardSuggestions = ({ fullname, id, profile,email ,follow}) => {
 
-    const dispatch = useDispatch();
     const [ button, setButton] =useState(false)
     const idToFollow = {"followMe": id}
     const myProfile = useSelector((state) => state.myProfileData)
     async function  followUnFollow() {
         await axios.put(`${process.env.REACT_APP_PUERTO}usuarios/follow/${myProfile[0].id}`, idToFollow , tokenUsuario())
-        dispatch(clearUsersState())
-        dispatch(getUsers())
+        setButton(true)
     } 
         
     
 
-    useEffect(() => {
-        follow.followers.includes(myProfile[0].id)===true?
-        setButton(true):setButton(false)
-    })
 
     return (
         <StyledCardSuggestions>
@@ -34,11 +27,8 @@ const CardSuggestions = ({ fullname, id, profile,email ,follow}) => {
                     <Link to={`/profile/${id}`} className="Link"><h3>{fullname}</h3></Link>
                     <p>{email}</p>
                 </div>
-                <div>
-                    {Object.keys(follow.followers).length?<div>si</div>:<div>no</div>}
                 </div>
-                </div>
-                { button===false?<button onClick={followUnFollow} className="btn"> seguir</button>:<button onClick={followUnFollow} className='btn2'>dejar de seguir</button>}
+                { button===false?<button onClick={followUnFollow} className="btn"> seguir</button>:<p>ahora sigues a {fullname}</p>}
             </div>
         </StyledCardSuggestions>
     )
