+++ Frontend/src/actions/index.js
@@ -5,13 +5,12 @@ import { CLEAR_PROFILE_STATE, CLEAR_USERS_STATE, GET_USER,  GET_MY_PROFILE, GET_
 export const getUsers = () => {
     return async (dispatch) => {
         try {
-            const usuarios = await axios.get(`${process.env.REACT_APP_PUERTO}usuarios/?myself=false`, tokenUsuario())
+            const usuarios = await axios.get(`${process.env.REACT_APP_PUERTO}usuarios/?myself=false&suggestions=true`, tokenUsuario())
             return dispatch({ type: GET_USER, payload: usuarios.data })
         } catch (err) {
             console.log(err)
         }
     }
-
 }
 
 export const clearUsersState = () => {
